.env: Initial commit
Custom values provided.

Note that at the time of this testing, this project does not work with a Specify version v7.9.4, therefore it is configured to use v7.9.1.1. It is unknown if a newer version of v7.9.4 will work.

ERROR in v7.9.4:
Collection Object Relations do not populate. Shown below it says "Loading..." in the "Relationship" field: